new layout

+++ app/bck/bm/datamanipulation/AdjustDataFrame.py
@@ -65,7 +65,7 @@ def encode_features_data_frame(model_id, data: DataFrame, column_type='F'):
     encoded_data = []
     data_types = data.dtypes
     for i in range(len(data_types)):
-        if data_types[i] != np.int64 and data_types[i] != np.float:
+        if data_types[i] != np.int64 and data_types[i] != float:
             col_name = columns_name[i]
             dummies = pd.get_dummies(data[[col_name]])
             dummies_columns = dummies.columns
@@ -97,9 +97,9 @@ def encode_labels_data_frame(model_id, data: DataFrame, column_type='L'):
         encoded_data = []
         data_types = data.dtypes
         for i in range(len(data_types)):
-            if data_types[i] != np.int64 and data_types[i] != np.float:
+            if data_types[i] != np.int64 and data_types[i] != float:
                 col_name = columns_name[i]
-                dummies = pd.get_dummies(data[[col_name]] if data_types[i] != np.float else round(data[[col_name]], 0))
+                dummies = pd.get_dummies(data[[col_name]] if data_types[i] != float else round(data[[col_name]], 0))
                 dummies_columns = dummies.columns
                 # encoded_data = encoded_data.append(dummies)
                 data = data.drop([col_name], axis=1)
